Master

+++ src/backend/database/migrations/2018_12_17_150648_create_projects_table.php
@@ -14,8 +14,8 @@ class CreateProjectsTable extends Migration
     public function up()
     {
         Schema::create('projects', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->unsignedBigInteger('owner_id');
+            $table->increments('id');
+            $table->unsignedInteger('owner_id');
             $table->string('title');
             $table->text('description');
             $table->text('notes')->nullable();
